[WIP] standardize expression construction
