+++ src/analysis/analysisVisitors/eventNameAnalysis/checkEventNamesAnalysis.test.ts
@@ -22,7 +22,8 @@ import {
 import CheckEventNamesAnalysis from "@/analysis/analysisVisitors/eventNameAnalysis/checkEventNamesAnalysis";
 import CustomEventEffect from "@/bricks/effects/customEvent";
 import { AnnotationType } from "@/types/annotationTypes";
-import { makeTemplateExpression } from "@/runtime/expressionCreators";
+
+import { makeTemplateExpression } from "@/utils/expressionUtils";
 
 describe("checkEventNamesAnalysis", () => {
   it("error on missing custom event", async () => {
