+++ src/bricks/transformers/brickFactory.test.ts
@@ -37,10 +37,11 @@ import { extraEmptyModStateContext } from "@/runtime/extendModVariableContext";
 import { setContext } from "@/testUtils/detectPageMock";
 import { uuidv4, validateRegistryId } from "@/types/helpers";
 import registerBuiltinBlocks from "@/bricks/registerBuiltinBlocks";
+
 import {
   makePipelineExpression,
   makeVariableExpression,
-} from "@/runtime/expressionCreators";
+} from "@/utils/expressionUtils";
 
 setContext("contentScript");
