+++ src/bricks/util.test.ts
@@ -18,8 +18,8 @@
 import { defaultBlockConfig, isOfficial } from "./util";
 import { type RegistryId } from "@/types/registryTypes";
 import IfElse from "./transformers/controlFlow/IfElse";
-import { EMPTY_PIPELINE } from "@/testUtils/testHelpers";
 import { type Schema } from "@/types/schemaTypes";
+import { EMPTY_PIPELINE } from "@/utils/expressionUtils";
 
 describe("isOfficial", () => {
   test("returns true for an official block", () => {
