+++ src/components/fields/schemaFields/widgets/TextWidget.tsx
@@ -35,13 +35,13 @@ import { isMustacheOnly } from "@/components/fields/fieldUtils";
 import { getToggleOptions } from "@/components/fields/schemaFields/getToggleOptions";
 import useUndo from "@/hooks/useUndo";
 import { isKeyStringField } from "@/components/fields/schemaFields/fieldTypeCheckers";
-import {
-  makeTemplateExpression,
-  makeVariableExpression,
-} from "@/runtime/expressionCreators";
 import { type Schema } from "@/types/schemaTypes";
 import { type TemplateEngine } from "@/types/runtimeTypes";
-import { isTemplateExpression } from "@/utils/expressionUtils";
+import {
+  isTemplateExpression,
+  makeTemplateExpression,
+  makeVariableExpression,
+} from "@/utils/expressionUtils";
 import { trimEndOnce } from "@/utils/stringUtils";
 
 function schemaSupportsTemplates(schema: Schema): boolean {
