+++ src/contrib/automationanywhere/BotOptions.test.tsx
@@ -23,7 +23,6 @@ import { Formik } from "formik";
 import { AUTOMATION_ANYWHERE_RUN_BOT_ID } from "@/contrib/automationanywhere/RunBot";
 import BotOptions from "@/contrib/automationanywhere/BotOptions";
 import useSanitizedIntegrationConfigFormikAdapter from "@/integrations/useSanitizedIntegrationConfigFormikAdapter";
-import { makeVariableExpression } from "@/runtime/expressionCreators";
 import registerDefaultWidgets from "@/components/fields/schemaFields/widgets/registerDefaultWidgets";
 import { type SanitizedConfig } from "@/integrations/integrationTypes";
 import { useAuthOptions } from "@/hooks/auth";
@@ -35,6 +34,7 @@ import {
 } from "@/testUtils/factories/integrationFactories";
 import { validateOutputKey } from "@/runtime/runtimeTypes";
 import { CONTROL_ROOM_TOKEN_INTEGRATION_ID } from "@/integrations/constants";
+import { makeVariableExpression } from "@/utils/expressionUtils";
 
 jest.mock("@/integrations/useSanitizedIntegrationConfigFormikAdapter", () => ({
   __esModule: true,
