+++ src/contrib/google/sheets/ui/LookupSpreadsheetOptions.tsx
@@ -25,8 +25,7 @@ import SchemaField from "@/components/fields/schemaFields/SchemaField";
 import { LOOKUP_SCHEMA } from "@/contrib/google/sheets/bricks/lookup";
 import { isEmpty } from "lodash";
 import { FormErrorContext } from "@/components/form/FormErrorContext";
-import { makeTemplateExpression } from "@/runtime/expressionCreators";
-import { isExpression } from "@/utils/expressionUtils";
+import { isExpression, makeTemplateExpression } from "@/utils/expressionUtils";
 import RequireGoogleSheet from "@/contrib/google/sheets/ui/RequireGoogleSheet";
 import { type SanitizedIntegrationConfig } from "@/integrations/integrationTypes";
 import { sheets } from "@/background/messenger/api";
