+++ src/contrib/google/sheets/ui/TabField.tsx
@@ -20,10 +20,9 @@ import { type SchemaFieldProps } from "@/components/fields/schemaFields/propType
 import SchemaField from "@/components/fields/schemaFields/SchemaField";
 import { isEmpty } from "lodash";
 import { useField } from "formik";
-import { makeTemplateExpression } from "@/runtime/expressionCreators";
 import { type Expression } from "@/types/runtimeTypes";
 import { type Schema } from "@/types/schemaTypes";
-import { isExpression } from "@/utils/expressionUtils";
+import { isExpression, makeTemplateExpression } from "@/utils/expressionUtils";
 import { type Spreadsheet } from "@/contrib/google/sheets/core/types";
 import useAsyncEffect from "use-async-effect";
