+++ src/pageEditor/exampleBlockConfigs.ts
@@ -22,10 +22,6 @@ import { CustomFormRenderer } from "@/bricks/renderers/customForm";
 import { createNewElement } from "@/components/documentBuilder/createNewElement";
 import DisplayTemporaryInfo from "@/bricks/transformers/temporaryInfo/DisplayTemporaryInfo";
 import { DocumentRenderer } from "@/bricks/renderers/document";
-import {
-  makePipelineExpression,
-  makeTemplateExpression,
-} from "@/runtime/expressionCreators";
 import { type BrickConfig } from "@/bricks/types";
 import { uuidv4 } from "@/types/helpers";
 import { defaultBlockConfig } from "@/bricks/util";
@@ -35,6 +31,10 @@ import { type RegistryId } from "@/types/registryTypes";
 import { type Schema } from "@/types/schemaTypes";
 import { JavaScriptTransformer } from "@/bricks/transformers/javascript";
 import { IdentityTransformer } from "@/bricks/transformers/identity";
+import {
+  makePipelineExpression,
+  makeTemplateExpression,
+} from "@/utils/expressionUtils";
 
 /**
  * Get a default block config for a block
