+++ src/pageEditor/starterBricks/pipelineMapping.test.ts
@@ -23,7 +23,6 @@ import {
   echoBrick,
   teapotBrick,
 } from "@/runtime/pipelineTests/pipelineTestHelpers";
-import { EMPTY_PIPELINE, toExpression } from "@/testUtils/testHelpers";
 import {
   normalizePipelineForEditor,
   omitEditorMetadata,
@@ -31,7 +30,11 @@ import {
 import blockRegistry from "@/bricks/registry";
 
 import { uuidSequence } from "@/testUtils/factories/stringFactories";
-import { isPipelineExpression } from "@/utils/expressionUtils";
+import {
+  EMPTY_PIPELINE,
+  isPipelineExpression,
+  toExpression,
+} from "@/utils/expressionUtils";
 import { type PipelineExpression } from "@/types/runtimeTypes";
 
 describe("normalizePipeline", () => {
